+++ ReleaseNotes.md
@@ -4,6 +4,7 @@
 
 - Bug fix: Fixed a bug when creating a tenant that had tenant roles
 - BREAKING CHANGE: GetAllRoleNamesAsync is now called GetRoleNamesForUsersAsync and takes the UserId
+- New Feature: Added GetRoleNamesForTenantsAsync to AuthTenantAdminService
 - Minor change: The UpdateRoleToPermissionsAsync method now allows you to change the RoleType